requirements.txt
flask>=2.0
nltk>=3.8
scikit-learn>=1.0
numpy>=1.21
pypdf>=3.0
rank_bm25>=0.2.2
faiss-cpu>=1.7.4
langdetect>=1.0.9
googletrans==4.0.0-rc1
pdfminer.six
# (optional: for better translation or offline) transformers[sentencepiece] sentencepiece

